requirements.txt
python-dotenv
numpy
pandas
langchain
langchain_community
langchain-huggingface 
huggingface_hub
Groq
langchain_groq
streamlit